requirements.txt
flask
flask-cors
python-dotenv
openai
gradio
requests
supabase
